Address 7DTD integration review feedback
