+++ Integrations/SevenDaysToDie/SevenDaysToDiePlayer.cs
@@ -1,19 +1,14 @@
+using Data.Models;
+using SharedLibraryCore.Database.Models;
+
 namespace Integrations.SevenDaysToDie;
 
-public sealed record SevenDaysToDiePlayer(
-    int Slot,
-    int EntityId,
-    long NetworkId,
-    string Name,
-    string Address,
-    int Ping,
-    int Level,
-    int ZombieKills,
-    int Deaths,
-    double PositionX,
-    double PositionY,
-    double PositionZ,
-    double RotationX,
-    double RotationY,
-    double RotationZ,
-    int Health);
+public sealed class SevenDaysToDiePlayer : EFClient
+{
+    public int EntityId { get; init; }
+    public int ZombieKills { get; init; }
+    public int PlayerDeaths { get; init; }
+    public Vector3 Position { get; init; } = new();
+    public Vector3 Rotation { get; init; } = new();
+    public int Health { get; init; }
+}
