+++ SharedLibraryCore/Configuration/ExternalWebConfiguration.cs
@@ -1,0 +1,15 @@
+using System;
+using SharedLibraryCore.Configuration.Attributes;
+
+namespace SharedLibraryCore.Configuration;
+
+public sealed class ExternalWebConfiguration
+{
+    public Uri Url { get; set; }
+
+    [ConfigurationOptional]
+    public string TokenName { get; set; }
+
+    [ConfigurationOptional]
+    public string TokenFile { get; set; }
+}
