+++ WebfrontCore/Components/Features/Servers/Components/ServerCard.razor
@@ -13,7 +13,7 @@
             <!-- 1. Game Tag -->
             <span
                 class="px-1.5 py-0.5 rounded bg-surface text-[10px] font-mono text-muted border border-line uppercase tracking-widest shrink-0 mr-3">
-                @(Model.Game == Data.Models.Reference.Game.D7D ? "7DTD" : Model.Game)
+                @Model.Game
             </span>
 
             <!-- 2. Server Name (High visual weight, expanded) -->
